# Optimize get_token by with inline methods