--- a/src/dateutil/parser/_parser.py
+++ b/src/dateutil/parser/_parser.py
@@ -95,6 +95,12 @@ class _timelex(object):
         token = None
         state = None
 
+        # Note:
+        # Inline optimizations
+        # nextchar.isalpha() := self.isword(nextchar)
+        # nextchar.isdigit() := self.isnum(nextchar)
+        # nextchar.isspace() := self.isspace(nextchar)
+
         while not self.eof:
             # We only realize that we've reached the end of a token when we
             # find a character that's not part of the current token - since
@@ -114,11 +120,11 @@ class _timelex(object):
                 # First character of the token - determines if we're starting
                 # to parse a word, a number or something else.
                 token = nextchar
-                if self.isword(nextchar):
+                if nextchar.isalpha():
                     state = 'a'
-                elif self.isnum(nextchar):
+                elif nextchar.isdigit():
                     state = '0'
-                elif self.isspace(nextchar):
+                elif nextchar.isspace():
                     token = ' '
                     break  # emit token
                 else:
@@ -127,7 +133,7 @@ class _timelex(object):
                 # If we've already started reading a word, we keep reading
                 # letters until we find something that's not part of a word.
                 seenletters = True
-                if self.isword(nextchar):
+                if nextchar.isalpha():
                     token += nextchar
                 elif nextchar == '.':
                     token += nextchar
@@ -138,7 +144,7 @@ class _timelex(object):
             elif state == '0':
                 # If we've already started reading a number, we keep reading
                 # numbers until we find something that doesn't fit.
-                if self.isnum(nextchar):
+                if nextchar.isdigit():
                     token += nextchar
                 elif nextchar == '.' or (nextchar == ',' and len(token) >= 2):
                     token += nextchar
@@ -150,9 +156,9 @@ class _timelex(object):
                 # If we've seen some letters and a dot separator, continue
                 # parsing, and the tokens will be broken up later.
                 seenletters = True
-                if nextchar == '.' or self.isword(nextchar):
+                if nextchar == '.' or nextchar.isalpha():
                     token += nextchar
-                elif self.isnum(nextchar) and token[-1] == '.':
+                elif nextchar.isdigit() and token[-1] == '.':
                     token += nextchar
                     state = '0.'
                 else:
@@ -161,9 +167,9 @@ class _timelex(object):
             elif state == '0.':
                 # If we've seen at least one dot separator, keep going, we'll
                 # break up the tokens later.
-                if nextchar == '.' or self.isnum(nextchar):
+                if nextchar == '.' or nextchar.isdigit():
                     token += nextchar
-                elif self.isword(nextchar) and token[-1] == '.':
+                elif nextchar.isalpha() and token[-1] == '.':
                     token += nextchar
                     state = 'a.'
                 else: